+++ src/config/db.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-console */
 import mongoose from 'mongoose';
 import { inspect } from 'util';
 
@@ -6,24 +5,30 @@ import CONSTANTS from './constants';
 
 mongoose.Promise = global.Promise;
 
-if (CONSTANTS.LOGGING.MONGO) {
-  mongoose.set('debug', (...rest) => {
-    Log[CONSTANTS.LOGGING.MONGO](inspect(rest));
-  });
-}
-(async () => {
+export default async () => {
+  // Mongo Debug
+  if (CONSTANTS.LOGGING.MONGO) {
+    mongoose.set('debug', (...rest) => {
+      Log[CONSTANTS.LOGGING.MONGO](inspect(rest));
+    });
+  }
+
+  // Connect
   try {
-    mongoose.connect(
+    await mongoose.connect(
       CONSTANTS.db.url,
       {},
     );
   } catch (err) {
-    mongoose.createConnection(CONSTANTS.db.url, {});
+    await mongoose.createConnection(CONSTANTS.db.url, {});
   }
 
+  // Open
   mongoose.connection.once('open', () => Log.info('MongoDB is running')).on('error', e => {
-    Log.error(JSON.stringify(e));
+    // Unknown if this ends up in main - therefore we log here
+    Log.error(e.stack);
+    throw e;
   });
-})();
+};
 
-export default mongoose;
+export { mongoose };