+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
-### 0.1.X
+### 0.1.15
 
-- [Fix] more dynamic connection whitelisting
+- [Changed] Filter performance for not-/voted
+- [Removed] Deprecated graphql field goverment [#438](https://github.com/demokratie-live/democracy-client/issues/438)
+- [Fixed] Search Button fix [#248](https://github.com/demokratie-live/democracy-client/issues/248)
 - [Fix] Remove id from Procedure.voteResults.PartyVotes subobject to fix unnecessary Push notifications
+- [Fix] more dynamic connection whitelisting
+- [Added] Catch all errors and log them
 
 ### 0.1.14